@@ -17,17 +17,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.1.6",
-    "@angular/cdk": "10.2.5",
-    "@angular/common": "10.1.6",
-    "@angular/compiler": "10.1.6",
-    "@angular/core": "10.1.6",
-    "@angular/forms": "10.1.6",
-    "@angular/localize": "10.1.6",
-    "@angular/material": "10.2.5",
-    "@angular/platform-browser": "10.1.6",
-    "@angular/platform-browser-dynamic": "10.1.6",
-    "@angular/router": "10.1.6",
+    "@angular/animations": "10.2.1",
+    "@angular/cdk": "10.2.6",
+    "@angular/common": "10.2.1",
+    "@angular/compiler": "10.2.1",
+    "@angular/core": "10.2.1",
+    "@angular/forms": "10.2.1",
+    "@angular/localize": "10.2.1",
+    "@angular/material": "10.2.6",
+    "@angular/platform-browser": "10.2.1",
+    "@angular/platform-browser-dynamic": "10.2.1",
+    "@angular/router": "10.2.1",
     "@auth0/angular-jwt": "5.0.1",
     "@fortawesome/angular-fontawesome": "0.7.0",
     "@fortawesome/fontawesome-svg-core": "1.2.32",
@@ -40,7 +40,7 @@
     "@ngrx/store-devtools": "10.0.1",
     "@swimlane/ngx-charts": "16.0.0",
     "bootstrap": "4.5.3",
-    "filepond": "4.21.1",
+    "filepond": "4.22.1",
     "filepond-plugin-file-validate-type": "1.2.5",
     "filepond-plugin-image-crop": "2.0.4",
     "filepond-plugin-image-exif-orientation": "1.0.9",
@@ -57,14 +57,14 @@
     "zone.js": "0.11.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1001.7",
-    "@angular/cli": "10.1.7",
-    "@angular/compiler-cli": "10.1.6",
-    "@angular/language-service": "10.1.6",
+    "@angular-devkit/build-angular": "0.1002.0",
+    "@angular/cli": "10.2.0",
+    "@angular/compiler-cli": "10.2.1",
+    "@angular/language-service": "10.2.1",
     "@ngrx/schematics": "10.0.1",
-    "@types/jasmine": "3.5.14",
+    "@types/jasmine": "3.6.0",
     "@types/jasminewd2": "2.0.8",
-    "@types/node": "13.13.26",
+    "@types/node": "13.13.30",
     "@types/socket.io": "2.1.11",
     "@types/ua-parser-js": "0.7.33",
     "codelyzer": "6.0.1",
@@ -77,7 +77,7 @@
     "karma-jasmine-html-reporter": "1.5.4",
     "prettier": "2.1.2",
     "protractor": "7.0.0",
-    "typescript": "4.0.3",
+    "typescript": "4.0.5",
     "ts-node": "9.0.0",
     "tslint": "6.1.3",
     "tslint-config-prettier": "1.18.0",
