@@ -1,16 +1,29 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
+
+import { PageQuery } from '@/shared/models';
+import { generatePageKey } from '@/shared/utils/generatePageKey';
+import { Page } from '@/shared/store';
 import * as fromDepartment from './department.reducer';
 import { Department } from '../models';
-import { PageQuery } from '@/shared/models';
 
 export const selectDepartmentsState = createFeatureSelector<fromDepartment.DepartmentsState>(
   fromDepartment.departmentsFeatureKey
 );
+export const selectDepartmentEntityState = createSelector(
+  selectDepartmentsState,
+  (departmentsState) => departmentsState.data
+);
+export const selectDepartmentPagesState = createSelector(
+  selectDepartmentsState,
+  (departmentsState) => departmentsState.pages
+);
 
 export const selectDepartmentById = (departmentId: number) =>
-  createSelector(selectDepartmentsState, (departmentsState) => departmentsState.entities[departmentId]);
+  createSelector(selectDepartmentEntityState, (departmentsState) => departmentsState.entities[departmentId]);
+export const selectDepartmentPageByKey = (pageQuery: PageQuery) =>
+  createSelector(selectDepartmentPagesState, (pagesState) => pagesState.entities[generatePageKey(pageQuery)]);
 
-export const selectAllDepartments = createSelector(selectDepartmentsState, fromDepartment.selectAll);
+export const selectAllDepartments = createSelector(selectDepartmentEntityState, fromDepartment.selectAll);
 
 export const selectDepartmentsLoading = createSelector(
   selectDepartmentsState,
@@ -23,22 +36,16 @@ export const selectDepartmentsPagesCount = createSelector(
 );
 
 export const selectDepartmentsTotalCount = createSelector(
-  selectDepartmentsState,
-  (departmentsState) => departmentsState.countAll
+  selectDepartmentPagesState,
+  (departmentsState) => departmentsState.resultsNum
 );
 
-export const selectDepartmentsPage = (page: PageQuery) =>
+export const selectDepartmentsOfPage = (pageQuery: PageQuery) =>
   createSelector(
     selectAllDepartments,
-    selectDepartmentsPagesCount,
-    (allDepartments: Department[], pagesCount: number) => {
-      if (!pagesCount) {
-        return [];
-      } else {
-        const start = page.pageIndex * page.pageSize,
-          end = start + page.pageSize;
-        return allDepartments.slice(start, end);
-      }
+    selectDepartmentPageByKey(pageQuery),
+    (allDepartments: Department[], page: Page) => {
+      return page ? page.dataIds.map((id) => allDepartments.find((department) => department.id === id)) : [];
     }
   );
 
