@@ -4,7 +4,7 @@ import { Plan } from '../../models';
 import { Store, select } from '@ngrx/store';
 import { AppState } from '@/core/reducers';
 import { PlansDataSource } from '../../services/plan.datasource';
-import { selectPlansLoading, selectPlansTotalCount } from '../../store/plan.selectors';
+import { selectPlanPageLoading, selectPlansTotalCount } from '../../store/plan.selectors';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { tap, takeUntil } from 'rxjs/operators';
@@ -41,7 +41,7 @@ import {
 })
 export class PlanListComponent implements AfterViewInit, OnDestroy {
   dataSource: PlansDataSource = new PlansDataSource(this.store$);
-  loading$: Observable<boolean> = this.store$.pipe(select(selectPlansLoading));
+  loading$: Observable<boolean> = this.store$.pipe(select(selectPlanPageLoading));
   resultsLength$: Observable<number> = this.store$.pipe(select(selectPlansTotalCount));
   canAddPlan$: Observable<boolean> = this.store$.pipe(select(isPrivileged(ADD_PRIVILEGES.PLAN)));
   canEditPlan$: Observable<boolean> = this.store$.pipe(select(isPrivileged(EDIT_PRIVILEGES.PLAN)));
@@ -79,7 +79,8 @@ export class PlanListComponent implements AfterViewInit, OnDestroy {
   ) {}
 
   ngAfterViewInit(): void {
-    merge(this.sort.sortChange, this.paginator.page)
+    const sort$ = this.sort.sortChange.pipe(tap(() => (this.paginator.pageIndex = 0)));
+    merge(sort$, this.paginator.page)
       .pipe(
         takeUntil(this.unsubscribe),
         tap(() => this.loadPlansPage())
@@ -93,7 +94,7 @@ export class PlanListComponent implements AfterViewInit, OnDestroy {
     const newPage: PageQuery = {
       pageIndex: this.paginator.pageIndex,
       pageSize: this.paginator.pageSize,
-      sortIndex: this.sort.active,
+      sortIndex: this.sort.active || COLUMN_NAMES.ID,
       sortDirection: this.sort.direction || SORT_DIRECTION.ASC
     };
 
