@@ -3,7 +3,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Plan } from '../models';
 import { User } from '@/modules/users';
-import { CollectionApiResponse, ApiResponse, ItemApiResponse } from '@/shared/models';
+import { CollectionApiResponse, ApiResponse, ItemApiResponse, PageQuery } from '@/shared/models';
 import { APIS } from '@/shared/constants';
 
 @Injectable()
@@ -14,12 +14,7 @@ export class PlanService {
     return this.http.post<ItemApiResponse<Plan>>(APIS.PLANS, this.formatBeforeSend(plan));
   }
 
-  getList(
-    pageIndex = 0,
-    pageSize = 5,
-    sortIndex = 'id',
-    sortDirection = 'asc'
-  ): Observable<CollectionApiResponse<Plan>> {
+  getList({ pageIndex, pageSize, sortIndex, sortDirection }: PageQuery): Observable<CollectionApiResponse<Plan>> {
     return this.http.get<CollectionApiResponse<Plan>>(APIS.PLANS, {
       params: new HttpParams()
         .set('pageIndex', pageIndex.toString())
@@ -28,13 +23,6 @@ export class PlanService {
         .set('sortDirection', sortDirection)
     });
   }
-
-  // getListByStage(stage: number, pageIndex = 0, pageSize = 5): Observable<Plan[]> {
-  //   const url = `${this.api}/stageList/${stage}`;
-  //   return this.http.get<Plan[]>(url, {
-  //     params: new HttpParams().set('pageIndex', pageIndex.toString()).set('pageSize', pageSize.toString())
-  //   });
-  // }
 
   getOne(id: number): Observable<ItemApiResponse<Plan>> {
     return this.http.get<ItemApiResponse<Plan>>(`${APIS.PLANS}/${id}`);
