@@ -3,7 +3,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Role } from '../models';
 import { User } from '@/modules/users/models';
-import { CollectionApiResponse, ItemApiResponse, ApiResponse } from '@/shared/models';
+import { CollectionApiResponse, ItemApiResponse, ApiResponse, PageQuery } from '@/shared/models';
 import { APIS } from '@/shared/constants';
 
 @Injectable()
@@ -26,12 +26,7 @@ export class RoleService {
     return this.http.delete<ApiResponse>(`${APIS.ROLES}/${id}`);
   }
 
-  getList(
-    pageIndex = 0,
-    pageSize = 5,
-    sortIndex = 'id',
-    sortDirection = 'asc'
-  ): Observable<CollectionApiResponse<Role>> {
+  getList({ pageIndex, pageSize, sortIndex, sortDirection }: PageQuery): Observable<CollectionApiResponse<Role>> {
     return this.http.get<CollectionApiResponse<Role>>(APIS.ROLES, {
       params: new HttpParams()
         .set('pageIndex', pageIndex.toString())
